// src/services/types.ts
export interface IFilterReport {
  currency?: string
  duration?: string
  startDate?: string
  endDate?: string
}

export interface ICollectionTrend {
  labels: string[]
  datasets: {
    label: string
    data: number[]
    maxBarThickness?: number
    backgroundColor?: string | string[]
    borderColor?: string
    tension?: number
    color?: string
  }[]
}

export enum DURATION {
  HOURLY = 'Hourly',
  DAILY = 'Daily',
  MONTHLY = 'Monthly'
}

export enum FilterMethod {
  Today = 'Today',
  ThisWeek = 'This Week',
  Last30Days = 'Last 30 Days',
  ThisMonth = 'This Month',
  Last90Days = 'Last 90 Days',
  ThisYear = 'This Year',
  CustomDate = 'Custom Date'
}